Grid view of posters are added to view on first open. (there is/was a concern where the view would only show the poster grid after rotating the image).
It seems likely, when the Adapter is instantiated in onCreateView, the MovieLibrary values are zero. Since the dataset contains no values, GridView does not know to create content views. Somewhere in the interaction between the Adapter data being updated and the GridView calling the custom adapter getView - Something doesn't expect more data. Perhaps something is set to null or 0 - either way. No posters are displayed on the GridView.